checkpoint: moves submodules over from "library/Base64CharacterEncodedByteSequence"
