+++ src/library/Base64CharacterEncodedByteSequence/exports.ts
@@ -1,3 +1,3 @@
 export {
   Sequence_Byte_Encoded_Base64,
-} from './definition.ts';
+} from '../Sequence/Byte/Encoded/Base64/definition.ts';
