+++ src/library/Sequence/Byte/Encoded/Base64/ParsingError.ts
